Move feed page to my/index
