+++ src/pages/my/index.tsx
@@ -1,15 +1,17 @@
 import { GetServerSideProps } from 'next';
-import { useEffect } from 'react';
-import { useRouter } from 'next/router';
+import { Heading } from '@adobe/react-spectrum';
+import { FormattedMessage as Msg } from 'react-intl';
 
 import { PageWithLayout } from '../../types';
 import { scaffold } from '../../utils/next';
 import UserHomeLayout from '../../components/layout/UserHomeLayout';
+import { ZetkinUser } from '../../interfaces/ZetkinUser';
 
 const scaffoldOptions = {
     localeScope: [
         'layout.userHome',
         'misc.publicHeader',
+        'pages.my.feed',
     ],
 };
 
@@ -19,20 +21,21 @@ export const getServerSideProps : GetServerSideProps = scaffold(async () => {
     };
 }, scaffoldOptions);
 
-const MyPage : PageWithLayout = () => {
+interface FeedProps {
+    user: ZetkinUser;
+}
 
-    const router = useRouter();
-
-    useEffect(() => {
-        router.push(`/my/feed`);
-    }, [router]);
+const Feed : PageWithLayout<FeedProps> = (props) => {
+    const { user } = props;
 
     return (
-        <></>
+        <Heading level={ 1 }>
+            <Msg id="pages.my.feed.welcome" values={{ userName: user.first_name }}/>
+        </Heading>
     );
 };
 
-MyPage.getLayout = function getLayout(page) {
+Feed.getLayout = function getLayout(page) {
     return (
         <UserHomeLayout>
             { page }
@@ -40,4 +43,4 @@ MyPage.getLayout = function getLayout(page) {
     );
 };
 
-export default MyPage;
+export default Feed;
